can wear an item using slots now

+++ server/priv/repo/migrations/20200515094007_create_item_physics.exs
@@ -6,10 +6,10 @@ defmodule Mud.Repo.Migrations.CreateItemPhysics do
       add(:id, :binary_id, primary_key: true)
       add(:item_id, references(:items, on_delete: :delete_all, type: :binary_id))
 
-      add(:length, :integer, default: 1)
-      add(:width, :integer, default: 1)
-      add(:height, :integer, default: 1)
-      add(:weight, :integer, default: 1)
+      add(:length, :numeric, default: 1)
+      add(:width, :numeric, default: 1)
+      add(:height, :numeric, default: 1)
+      add(:weight, :numeric, default: 1)
     end
 
     create(index(:item_physics, [:length]))
